Revert "refactor: 💡 clientId to long + test with 128bit key"
This reverts commit 933bd46b304627db42c990703f61a773eee5db9d.

+++ MovieCrew.API.Test/Integration/Authentication/TokenEndpointsTest.cs
@@ -68,7 +68,7 @@ public class TokenEndpointsTest
     public async Task AuthenticationShouldReturnForbiddenIfNotValidClient()
     {
         // Arrange
-        _authenticationService.Setup(x => x.Authenticate(It.IsAny<long>(), It.IsAny<string>()))
+        _authenticationService.Setup(x => x.Authenticate(It.IsAny<int>(), It.IsAny<string>()))
             .ThrowsAsync(new AuthenticationException("Invalid client"));
         _client.DefaultRequestHeaders.Add("ApiKey", "test");
